move optimal_clip into the IF
