+++ safe/impact_functions/base.py
@@ -35,9 +35,10 @@ from safe.impact_functions.impact_function_metadata import \
 from safe.common.exceptions import (
     InvalidExtentError,
     FunctionParametersError,
+    CallGDALError,
     NoValidLayerError,
     InsufficientOverlapError)
-from safe.common.signals import send_static_message
+from safe.common.signals import send_static_message, send_dynamic_message
 from safe.postprocessors.postprocessor_factory import (
     get_postprocessors,
     get_postprocessor_human_name)
@@ -53,7 +54,7 @@ from safe.utilities.gis import (
     get_wgs84_resolution,
     viewport_geo_array,
     get_optimal_extent)
-from safe.utilities.clipper import adjust_clip_extent
+from safe.utilities.clipper import adjust_clip_extent, clip_layer
 from safe.storage.safe_layer import SafeLayer
 from safe.storage.utilities import (
     buffered_bounding_box as get_buffered_extent,
@@ -102,7 +103,7 @@ class ImpactFunction(object):
         self._callback = self.console_progress_callback
         # Set the default parameters
         self._parameters = self._metadata.parameters()
-        # Clip parameters
+        # Clip parameters - Read only
         self._clip_parameters = None
         # Layer representing hazard e.g. flood
         self._hazard = None
@@ -126,6 +127,8 @@ class ImpactFunction(object):
         self._not_affected_value = 'Not Affected'
         # Store provenances
         self._provenances = Provenance()
+        # Clip hard
+        self._clip_hard = None
         # Start time
         self._start_time = None
 
@@ -377,6 +380,17 @@ class ImpactFunction(object):
         :type parameters: dict
         """
         self._parameters = parameters
+
+    @property
+    def clip_hard(self):
+        if self._clip_hard is None:
+            return False
+        else:
+            return self._clip_hard
+
+    @clip_hard.setter
+    def clip_hard(self, clip_hard):
+        self._clip_hard = clip_hard
 
     @property
     def impact(self):
@@ -879,3 +893,106 @@ class ImpactFunction(object):
             'adjusted_geo_extent': adjusted_geo_extent,
             'cell_size': cell_size
         }
+
+    def optimal_clip(self):
+        """ A helper function to perform an optimal clip of the input data.
+        Optimal extent should be considered as the intersection between
+        the three inputs. The InaSAFE library will perform various checks
+        to ensure that the extent is tenable, includes data from both
+        etc.
+
+        The result of this function will be two layers which are
+        clipped and re-sampled if needed, and in the EPSG:4326 geographic
+        coordinate reference system.
+
+        :returns: The clipped hazard and exposure layers.
+        :rtype: (QgsMapLayer, QgsMapLayer)
+        """
+
+        # Get the hazard and exposure layers selected in the combos
+        # and other related parameters needed for clipping.
+        try:
+            extra_exposure_keywords = self.clip_parameters[
+                'extra_exposure_keywords']
+            adjusted_geo_extent = self.clip_parameters['adjusted_geo_extent']
+            cell_size = self.clip_parameters['cell_size']
+        except:
+            raise
+        # Find out what clipping behaviour we have - see #2210
+        settings = QSettings()
+        mode = settings.value(
+            'inasafe/analysis_extents_mode',
+            'HazardExposureView')
+        detail = None
+        if mode == 'HazardExposureView':
+            detail = tr(
+                'Resampling and clipping the hazard layer to match the '
+                'intersection of the exposure layer and the current view '
+                'extents.')
+        elif mode == 'HazardExposure':
+            detail = tr(
+                'Resampling and clipping the hazard layer to match the '
+                'intersection of the exposure layer extents.')
+        elif mode == 'HazardExposureBookmark':
+            detail = tr(
+                'Resampling and clipping the hazard layer to match the '
+                'bookmarked extents.')
+        elif mode == 'HazardExposureBoundingBox':
+            detail = tr(
+                'Resampling and clipping the hazard layer to match the '
+                'intersection of your preferred analysis area.')
+        # Make sure that we have EPSG:4326 versions of the input layers
+        # that are clipped and (in the case of two raster inputs) resampled to
+        # the best resolution.
+        title = tr('Preparing hazard data')
+
+        message = m.Message(
+            m.Heading(title, **PROGRESS_UPDATE_STYLE),
+            m.Paragraph(detail))
+        send_dynamic_message(self, message)
+        try:
+            clipped_hazard = clip_layer(
+                layer=self.hazard.qgis_layer(),
+                extent=adjusted_geo_extent,
+                cell_size=cell_size,
+                hard_clip_flag=self.clip_hard)
+        except CallGDALError, e:
+            raise e
+        except IOError, e:
+            raise e
+
+        title = tr('Preparing exposure data')
+        # Find out what clipping behaviour we have - see #2210
+        settings = QSettings()
+        mode = settings.value(
+            'inasafe/analysis_extents_mode',
+            'HazardExposureView')
+        if mode == 'HazardExposureView':
+            detail = tr(
+                'Resampling and clipping the exposure layer to match '
+                'the intersection of the hazard layer and the current view '
+                'extents.')
+        elif mode == 'HazardExposure':
+            detail = tr(
+                'Resampling and clipping the exposure layer to match '
+                'the intersection of the hazard layer extents.')
+        elif mode == 'HazardExposureBookmark':
+            detail = tr(
+                'Resampling and clipping the exposure layer to match '
+                'the bookmarked extents.')
+        elif mode == 'HazardExposureBoundingBox':
+            detail = tr(
+                'Resampling and clipping the exposure layer to match '
+                'the intersection of your preferred analysis area.')
+        message = m.Message(
+            m.Heading(title, **PROGRESS_UPDATE_STYLE),
+            m.Paragraph(detail))
+        send_dynamic_message(self, message)
+
+        clipped_exposure = clip_layer(
+            layer=self.exposure.qgis_layer(),
+            extent=adjusted_geo_extent,
+            cell_size=cell_size,
+            extra_keywords=extra_exposure_keywords,
+            hard_clip_flag=self.clip_hard)
+        return clipped_hazard, clipped_exposure
